refactor(utils): make rw_futures_util a standalone crate
