+++ src/batch/Cargo.toml
@@ -37,6 +37,7 @@ risingwave_pb = { workspace = true }
 risingwave_rpc_client = { workspace = true }
 risingwave_source = { workspace = true }
 risingwave_storage = { workspace = true }
+rw_futures_util = { workspace = true }
 scopeguard = "1"
 serde_json = "1"
 thiserror = "1"
