+++ src/common/src/util/mod.rs
@@ -21,7 +21,6 @@ pub mod compress;
 pub mod deployment;
 pub mod env_var;
 pub mod epoch;
-mod future_utils;
 pub mod hash_util;
 pub mod iter_util;
 pub mod memcmp_encoding;
@@ -42,9 +41,5 @@ pub mod tracing;
 pub mod value_encoding;
 pub mod worker_util;
 
-pub use future_utils::{
-    await_future_with_monitor_error_stream, drop_either_future, pending_on_none, select_all,
-    RwFutureExt, RwTryStreamExt,
-};
 #[macro_use]
 pub mod match_util;
