+++ src/frontend/Cargo.toml
@@ -67,6 +67,7 @@ risingwave_sqlparser = { workspace = true }
 risingwave_storage = { workspace = true }
 risingwave_udf = { workspace = true }
 risingwave_variables = { workspace = true }
+rw_futures_util = { workspace = true }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 sha2 = "0.10.7"
