+++ src/meta/src/barrier/rpc.rs
@@ -24,7 +24,7 @@ use futures::stream::FuturesUnordered;
 use futures::{FutureExt, StreamExt};
 use itertools::Itertools;
 use risingwave_common::bail;
-use risingwave_common::util::pending_on_none;
+use rw_futures_util::pending_on_none;
 use risingwave_common::util::tracing::TracingContext;
 use risingwave_pb::stream_plan::{Barrier, BarrierMutation};
 use risingwave_pb::stream_service::{BarrierCompleteRequest, InjectBarrierRequest};
