+++ src/rpc_client/Cargo.toml
@@ -29,6 +29,7 @@ risingwave_common = { workspace = true }
 risingwave_error = { workspace = true }
 risingwave_hummock_sdk = { workspace = true }
 risingwave_pb = { workspace = true }
+rw_futures_util = { workspace = true }
 static_assertions = "1"
 thiserror = "1"
 thiserror-ext = { workspace = true }
