+++ src/stream/Cargo.toml
@@ -54,6 +54,7 @@ risingwave_pb = { workspace = true }
 risingwave_rpc_client = { workspace = true }
 risingwave_source = { workspace = true }
 risingwave_storage = { workspace = true }
+rw_futures_util = { workspace = true }
 serde_json = "1"
 smallvec = "1"
 static_assertions = "1"
