+++ src/utils/futures_util/src/misc.rs
@@ -12,15 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod buffered_with_fence;
-
-use std::future::pending;
+use std::future::Future;
 use std::pin::{pin, Pin};
 
-pub use buffered_with_fence::*;
-use futures::future::{select, Either};
+use futures::future::{pending, select, Either};
 use futures::stream::Peekable;
-use futures::{Future, FutureExt, Stream, StreamExt};
+use futures::{FutureExt, Stream, StreamExt};
 
 /// Convert a list of streams into a [`Stream`] of results from the streams.
 pub fn select_all<S: Stream + Unpin>(
